+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] (alpha)
 
+### Fixed
+
+- Check if the user specified for an exchange backup operation has a mailbox.
+
 
 ## [v0.1.0] (alpha) - 2023-01-13